refactor: moved POST /users from UserController to /auth/register in AuthController
