+++ backend/account-service/src/main/java/com/ita/challenges/account/infrastructure/adapter/web/in/UserController.java
@@ -1,9 +1,6 @@
 package com.ita.challenges.account.infrastructure.adapter.web.in;
 
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -11,9 +8,4 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/account/users")
 @CrossOrigin(origins = "http://localhost:4200")
 public class UserController {
-
-    @PostMapping
-    public ResponseEntity<Void> create() {
-        return ResponseEntity.status(HttpStatus.OK).build();
-    }
 }
