+++ backend/account-service/src/test/java/com/ita/challenges/account/infrastructure/adapter/web/in/AuthControllerTest.java
@@ -5,6 +5,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.test.web.servlet.MockMvc;
 
+import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.oauth2Login;
@@ -16,6 +18,15 @@ class AuthControllerTest {
 
     @Autowired
     private MockMvc mockMvc;
+
+    @Test
+    void create_shouldReturn200() throws Exception {
+        mockMvc.perform(post("/api/account/auth/register")
+                        .with(oauth2Login())
+                        .with(csrf())
+                )
+                .andExpect(status().isOk());
+    }
 
     @Test
     void me_whenAuthenticated_returnsUsernameAndAvatarUrl() throws Exception {
