+++ backend/account-service/src/test/java/com/ita/challenges/account/infrastructure/adapter/web/in/UserControllerTest.java
@@ -1,27 +1,7 @@
 package com.ita.challenges.account.infrastructure.adapter.web.in;
 
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.test.web.servlet.MockMvc;
-
-import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
-import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.oauth2Login;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @WebMvcTest(UserController.class)
 class UserControllerTest {
-
-    @Autowired
-    private MockMvc mockMvc;
-
-    @Test
-    void create_shouldReturn200() throws Exception {
-        mockMvc.perform(post("/api/account/users")
-                        .with(oauth2Login())
-                        .with(csrf())
-                )
-                .andExpect(status().isOk());
-    }
 }
